added currentWatcher
I attempted to add the currentWatcher to the collector. I also chenged the currentSupplier in the currentWatcher class from int to double because that's the only way it worked with the motor. I also just randomly guessed numbers for the max time and maxCurrent as I have no idea what to put there