Collector (#1)
* started constants, made enum

* added neutral mode and inverted mode value

* fixed small line mistakes

* fixed TalonSRX config

* added voltage compensation

* removed extra new line

* removed extra new line in subsystem

* updated code acording to review

* removed extra new line

* added motor to subsystem

* added note

* added stop and atempted setCollectorStateCommand

btw, I doubt motor.setIntegralAccumulator is right but it's my best guess so far.

* finished subsystem

* fixed

* added javadocs

* removed extra new lines

* fixed javadocs

* updated command

* added currentWatcher

I attempted to add the currentWatcher to the collector. I also chenged the currentSupplier in the currentWatcher class from int to double because that's the only way it worked with the motor. I also just randomly guessed numbers for the max time and maxCurrent as I have no idea what to put there

* fixed implementation of CurrentWatcher

* updated numbers

* corrected implementation of currentWatcher

* made startCurrentWatcher

* changed neutralmode to coast